Make MaterialPropertyBlock accessible to use for synchronization
