+++ samples/SpectatorView.Example.Unity/Assets/Scenes/SpectatorView.HoloLens.unity.meta
