+++ samples/SpectatorView.Example.Unity/Assets/Scripts/Demo/ChangeColor.cs
@@ -10,7 +10,7 @@ namespace Microsoft.MixedReality.SpectatorView
     internal class ChangeColor : MonoBehaviour
     {
         [SerializeField]
-        private Material material = null;
+        private Renderer targetRenderer = null;
 
         void Update()
         {
@@ -18,7 +18,7 @@ namespace Microsoft.MixedReality.SpectatorView
             float g = (1.0f + Mathf.Sin(2 * Time.time)) / 2.0f;
             float b = (1.0f + Mathf.Sin(3 * Time.time)) / 2.0f;
             float alpha = (1.0f + Mathf.Sin(4 * Time.time)) / 2.0f;
-            material.color = new Color(r, g, b, alpha);
+            targetRenderer.material.color = new Color(r, g, b, alpha);
         }
     }
 }
