+++ samples/SpectatorView.Example.Unity/Assets/Scripts/Demo/ChangeColor.cs.meta
