+++ samples/SpectatorView.Example.Unity/Assets/Scripts/Demo/ChangeMaterialProperty.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 080592ca1eeef704dac0a5d91bc664c3
+guid: d1fb2846b447a614aa733ea62dab9bfa
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
