+++ samples/SpectatorView.Example.Unity/Assets/Scripts/Demo/Spin.cs
@@ -1,0 +1,25 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+
+using UnityEngine;
+
+namespace Microsoft.MixedReality.SpectatorView
+{
+    internal class Spin : MonoBehaviour
+    {
+        /// <summary>
+        /// SpectatorView instance
+        /// </summary>
+        [Tooltip("SpectatorView instance")]
+        [SerializeField]
+        private SpectatorView spectatorView = null;
+
+        void Update()
+        {
+            if (spectatorView != null && spectatorView.Role == Role.User)
+            {
+                gameObject.transform.localRotation = Quaternion.Euler(0, 100 * Time.time, 0);
+            }
+        }
+    }
+}
