+++ samples/SpectatorView.Example.Unity/Assets/Scripts/Demo/Spin.cs.meta
